+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+12 February
+-----------
+
+- Implemented the Reinforce algorithm (simple policy gradient)
+- Able to solve the CartPole-v1 environment in 400 steps
+
 23 October
 ----------